implement jobs applications

+++ django_crud/articles/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Article
 
 # articles의 메인페이지고, article list를 보여준다.
@@ -12,28 +12,62 @@ def index(request):
     # render 순서를 잘 지켜야 한다.
 
 def detail(request, article_pk):
-    article = Article.objects.get(pk=article_pk)
+    article = get_object_or_404(Article, pk=article_pk)
     context = {
         'article': article
         }    
     return render(request, 'articles/detail.html', context)
 
-def new(request):
-    return render(request, 'articles/new.html')
-
+# 입력페이지 제공
+# GET /articles/create/
+# def new(request):
+        # return render(request, 'articles/new.html')
+    
 # new.html에서 사용자가 입력한 정보가 여기로 보내진다.
 # /articles/new/의 new.html 의 form에서 전달받은 데이터들
+# POST /articles/create/
 def create(request):
-    title = request.GET.get('title')
-    content = request.GET.get('content')
-    article = Article() # 생성한 스키마의 새로운 인스턴스 생성
-    article.title = title # 정보를 할당시켜준다.
-    article.content = content
-    article.save() # 저장
-    # redirect를 사용해서 index로 이동
-    return redirect(f'/articles/detail/{article.pk}/')
+    # 만약 GET 요청으로 들어오면 html 페이지 rendering
+    # 아니라면 (POST일 경우) 사용자 데이터 받아서 article 생성
+    if request.method == 'POST':
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+        article = Article() # 생성한 스키마의 새로운 인스턴스 생성
+        article.title = title # 정보를 할당시켜준다.
+        article.content = content
+        article.save() # 저장
+        # redirect를 사용해서 index로 이동
+        return redirect('articles:detail', article.pk)
+    else:
+        return render(request, 'articles/create.html')
+# redirect(f'/articles/detail/{article.pk}/')
 
 def delete(request, article_pk):
-    article = Article.objects.get(pk=article_pk)
-    article.delete()
-    return redirect(f'/articles/')
+    article = get_object_or_404(Article, pk=article_pk)
+    # article = Article.objects.get(pk=article_pk)
+    if request.method == 'POST': 
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return redirect('articles:detail', article.pk)
+# redirect(f'/articles/')
+
+# /articles/5/update/
+def update(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    # POST : 실제 Update 로직 수행
+    if request.method == 'POST':
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+
+        article.title = title # 정보를 할당시켜준다.
+        article.content = content
+        article.save() # 저장
+        return redirect('articles:detail', article.pk)
+
+    # GET : Update 를 하기위한 Form을 제공하는 페이지
+    else: # POST가 아니라 GET이라면,
+        context = {
+            'article' : article
+        }
+        return render(request, 'articles/update.html', context)
